+++ CHANGES.md
@@ -1,3 +1,9 @@
+# 11 Feb 2014
+
+* Add Cache-Control Header into response
+* Add ETag Header into response
+* Upgrade Go version and compilation
+
 # 28 june 2013
 
 * Add multiplatform compilation